+++ src/model/index.ts
@@ -8,3 +8,4 @@ export { Region } from './Region';
 export { PseudoState } from './PseudoState';
 export { Transition } from './Transition';
 export { TransitionKind } from './TransitionKind';
+export { TransitionActivation, ExternalTransitionActivation, LocalTransitionActivation, InternalTransitionActivation } from './TransitionActivation';